examples/04-structured-output/requirements.txt
djangosdk>=0.1.1
python-decouple>=3.8
